requirements.txt
setuptools==40.8.0
python_string_utils==0.6.0
requests==2.22.0
Pillow==6.1.0
